Moved test case to a new file
